requirements-scientific.txt
# Scientific Computing Requirements for Obvivlorum Research Components
# Install with: pip install -r requirements-scientific.txt

# Core Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
numba>=0.56.0
matplotlib>=3.5.0
networkx>=2.6.0

# Quantum Computing (Optional but Recommended)
qiskit>=0.44.0
pennylane>=0.32.0
cirq>=1.0.0

# Machine Learning and Neural Networks
scikit-learn>=1.0.0
torch>=1.12.0
tensorflow>=2.10.0

# Performance and Optimization
cython>=0.29.0
numexpr>=2.8.0
bottleneck>=1.3.0

# Data Analysis and Visualization
pandas>=1.4.0
seaborn>=0.11.0
plotly>=5.10.0

# Statistical Analysis
statsmodels>=0.13.0
pymc>=5.0.0

# Memory Profiling and Performance
memory-profiler>=0.60.0
psutil>=5.8.0
py-spy>=0.3.0

# Parallel Processing
joblib>=1.1.0
dask>=2022.8.0
ray>=2.0.0

# Graph Theory and Complex Networks
graph-tool>=2.45  # Optional, system-dependent
igraph>=0.9.0

# Information Theory
dit>=1.4.0  # Discrete Information Theory

# Optimization
cvxpy>=1.2.0
optuna>=3.0.0

# Testing and Benchmarking
pytest-benchmark>=4.0.0
hypothesis>=6.0.0

# Documentation
sphinx>=5.0.0
sphinx-autodoc-typehints>=1.19.0
myst-parser>=0.18.0

# Jupyter and Interactive Development
jupyter>=1.0.0
ipywidgets>=8.0.0
jupyterlab>=3.4.0